Add locations module and move Locations classes
