+++ seacharts/layers/layer.py
@@ -1,7 +1,7 @@
 from __future__ import annotations
 
 from abc import ABC
-from dataclasses import dataclass, field
+from dataclasses import dataclass
 
 from shapely import geometry as geo
 
@@ -48,23 +48,3 @@ class MultiDepth:
     @property
     def depth(self) -> None:
         raise AttributeError("Multi-depth shapes have no single depth.")
-
-
-@dataclass
-class Locations(Layer, ABC):
-    geometry: geo.MultiPoint = field(default_factory=geo.MultiPoint)
-
-
-@dataclass
-class ZeroDepthLocations(Locations, ZeroDepth, ABC):
-    pass
-
-
-@dataclass
-class SingleDepthLocations(Locations, SingleDepth, ABC):
-    pass
-
-
-@dataclass
-class MultiDepthLocations(Locations, MultiDepth, ABC):
-    pass
